requirements.txt
alembic==1.3.0
amqp==2.5.1
Babel==2.7.0
billiard==3.6.1.0
blinker==1.4
celery==4.3.0
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.5
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
mysql-connector-python==8.0.18
protobuf==3.10.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
redis==3.3.11
six==1.12.0
SQLAlchemy==1.3.10
vine==1.3.0
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
